requirements.txt
pandas==1.2.5
sqlalchemy==1.3.23
psycopg2-binary==2.8.6
python-dotenv==0.15.0
